:construction: Search for vivaldi dirs near the entered path
When entering a path to a vivaldi dir and unfocusing the field, the path is checked to see if it contains one or more version folders (= is a valid vivaldi dir).

Now, if the check fails, the parent and child folders will also be checked for a vivaldi dir. If one or more are found, they are all added to the vivaldi dir list.

Still slightly buggy, and doesn't yet work for folders chosen using "Browse".